requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
aiosqlite==0.20.0
